requirements.txt
flask
ntplib
sqlalchemy
requests
fake-useragent
PyMySQL
PyYAML
psutil
